# arks_servers/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Acks 博客项目 API 接口文档",
        "title": "web接口bolg",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8888",
    "paths": {
        "/blog/captcha": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "授权"
                ],
                "summary": "创建验证码",
                "responses": {
                    "100": {
                        "description": "code\":10000,\"data\":base64,\"msg\":\"验证码创建成功\"}",
                        "schema": {
                            "$ref": "#/definitions/utils.Result"
                        }
                    },
                    "104": {
                        "description": "code\":10001,\"data\":base64,\"msg\":\"服务器端错误\"}",
                        "schema": {
                            "$ref": "#/definitions/utils.Result"
                        }
                    }
                }
            }
        },
        "/blog/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "授权"
                ],
                "summary": "登录",
                "responses": {
                    "100": {
                        "description": "Continue",
                        "schema": {
                            "$ref": "#/definitions/utils.Result"
                        }
                    },
                    "104": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/utils.Result"
                        }
                    }
                }
            }
        },
        "/blog/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "授权"
                ],
                "summary": "注册",
                "responses": {
                    "100": {
                        "description": "Continue",
                        "schema": {
                            "$ref": "#/definitions/utils.Result"
                        }
                    },
                    "104": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/utils.Result"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "utils.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "type": "object"
                },
                "msg": {
                    "description": "提示",
                    "type": "string"
                }
            }
        }
    }
}